+++ source/DSCResources/MSFT_SqlDatabase/MSFT_SqlDatabase.psm1
@@ -29,7 +29,7 @@ $script:supportedCompatibilityLevels = @{
     The name of database to be created or dropped.
 
     .PARAMETER ServerName
-    The host name of the SQL Server to be configured.
+    The host name of the SQL Server to be configured. Defaults to $env:COMPUTERNAME.
 
     .PARAMETER InstanceName
     The name of the SQL instance to be configured.
@@ -64,10 +64,10 @@ function Get-TargetResource
         [System.String]
         $Name,
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [ValidateNotNullOrEmpty()]
         [System.String]
-        $ServerName,
+        $ServerName = $env:COMPUTERNAME,
 
         [Parameter(Mandatory = $true)]
         [ValidateNotNullOrEmpty()]
@@ -146,7 +146,7 @@ function Get-TargetResource
     The name of database to be created or dropped.
 
     .PARAMETER ServerName
-    The host name of the SQL Server to be configured.
+    The host name of the SQL Server to be configured. Defaults to $env:COMPUTERNAME.
 
     .PARAMETER InstanceName
     The name of the SQL instance to be configured.
@@ -179,10 +179,10 @@ function Set-TargetResource
         [System.String]
         $Name,
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [ValidateNotNullOrEmpty()]
         [System.String]
-        $ServerName,
+        $ServerName = $env:COMPUTERNAME,
 
         [Parameter(Mandatory = $true)]
         [ValidateNotNullOrEmpty()]
@@ -325,7 +325,7 @@ function Set-TargetResource
     The name of database to be created or dropped.
 
     .PARAMETER ServerName
-    The host name of the SQL Server to be configured.
+    The host name of the SQL Server to be configured. Defaults to $env:COMPUTERNAME.
 
     .PARAMETER InstanceName
     The name of the SQL instance to be configured.
@@ -359,10 +359,10 @@ function Test-TargetResource
         [System.String]
         $Name,
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [ValidateNotNullOrEmpty()]
         [System.String]
-        $ServerName,
+        $ServerName = $env:COMPUTERNAME,
 
         [Parameter(Mandatory = $true)]
         [ValidateNotNullOrEmpty()]